data
Novel
From Wikipedia, the free encyclopedia
Jump to navigationJump to search
For other uses, see Novel (disambiguation).
Not to be confused with Novell.

This article needs additional citations for verification. Please help improve this article by adding citations to reliable sources. Unsourced material may be challenged and removed.
Find sources: "Novel" – news · newspapers · books · scholar · JSTOR (April 2018) (Learn how and when to remove this template message)
Literature
Liji2 no bg.png
Major forms
NovelPoetryDramaShort storyNovella
Genres
AdventureComedyDramaEpicEroticNonsenseLyricMythopoeiaRogueRomanceSatireSpeculative fictionTragedyTragicomedy
Media
Performance PlayBooks
Techniques
ProsePoetry
History and lists
History modern
OutlineGlossary of terms
BooksWritersLiterary awards poetry
Discussion
CriticismTheory (critical theory)SociologyMagazines
Books-aj.svg aj ashton 01.svg Literature portal
vte
A novel is a relatively long work of narrative fiction, typically written in prose and published as a book. The present English word for a long work of prose fiction derives from the Italian: novella for "new", "news", or "short story of something new", itself from the Latin: novella, a singular noun use of the neuter plural of novellus, diminutive of novus, meaning "new".[1]

Some novelists, including Nathaniel Hawthorne,[2] Herman Melville,[3] Ann Radcliffe,[4] John Cowper Powys,[5] preferred the term "romance" to describe their novels.

According to Margaret Doody, the novel has "a continuous and comprehensive history of about two thousand years", with its origins in the Ancient Greek and Roman novel, in Chivalric romance, and in the tradition of the Italian renaissance novella.[6] The ancient romance form was revived by Romanticism, especially the historical romances of Walter Scott and the Gothic novel.[7] Some, including M. H. Abrams and Walter Scott, have argued that a novel is a fiction narrative that displays a realistic depiction of the state of a society, while the romance encompasses any fictitious narrative that emphasizes marvellous or uncommon incidents.[8][9][10]

Works of fiction that include marvellous or uncommon incidents are also novels, including The Lord of The Rings,[11] To Kill a Mockingbird,[12] and Frankenstein.[13] "Romances" are works of fiction whose main emphasis is on marvellous or unusual incidents, and should not be confused with the romance novel, a type of genre fiction that focuses on romantic love.

Murasaki Shikibu's Tale of Genji, an early 11th-century Japanese text, has sometimes been described as the world's first novel, but there is considerable debate over this — there were certainly long fictional works that preceded it. Spread of printed books in China led to the appearance of classical Chinese novels by the Ming dynasty (1368–1644). An early example from Europe was written in Muslim Spain by the Sufi writer Ibn Tufayl entitled Hayy ibn Yaqdhan.[14] Later developments occurred after the invention of the printing press. Miguel de Cervantes, author of Don Quixote (the first part of which was published in 1605), is frequently cited as the first significant European novelist of the modern era.[15] Literary historian Ian Watt, in The Rise of the Novel (1957), argued that the modern novel was born in the early 18th century.


Contents
1	Defining the genre
1.1	Fictional narrative
1.2	Literary prose
1.3	Experience of intimacy
1.4	Length
2	Early novels
3	Medieval period 1100–1500
3.1	Chivalric romances
3.2	The novella
4	Renaissance period: 1500–1700
4.1	Chapbooks
4.2	Heroic romances
4.3	Satirical romances
4.4	Histories
4.5	Cervantes and the modern novel
5	18th-century novels
5.1	Philosophical novel
5.2	The romance genre in the 18th century
5.3	The sentimental novel
5.4	The social context of the 18th century novel
5.4.1	Changing cultural status
5.4.2	The acceptance of novels as literature
6	19th-century novels
6.1	Romanticism
6.2	The Victorian period: 1837–1901
7	The 20th century and later
7.1	Modernism and post-modernism
7.2	Genre fiction
8	See also
9	References
10	Further reading
11	External links
Defining the genre

Madame de Pompadour spending her afternoon with a book (François Boucher, 1756)
A novel is a long, fictional narrative which describes intimate human experiences. The novel in the modern era usually makes use of a literary prose style. The development of the prose novel at this time was encouraged by innovations in printing, and the introduction of cheap paper in the 15th century.

Fictional narrative
Fictionality is most commonly cited as distinguishing novels from historiography. However this can be a problematic criterion. Throughout the early modern period authors of historical narratives would often include inventions rooted in traditional beliefs in order to embellish a passage of text or add credibility to an opinion. Historians would also invent and compose speeches for didactic purposes. Novels can, on the other hand, depict the social, political and personal realities of a place and period with clarity and detail not found in works of history.

Literary prose
While prose rather than verse became the standard of the modern novel, the ancestors of the modern European novel include verse epics in the Romance language of southern France, especially those by Chrétien de Troyes (late 12th century), and in Middle English (Geoffrey Chaucer's (c. 1343 – 1400) The Canterbury Tales).[16] Even in the 19th century, fictional narratives in verse, such as Lord Byron's Don Juan (1824), Alexander Pushkin's Yevgeniy Onegin (1833), and Elizabeth Barrett Browning's Aurora Leigh (1856), competed with prose novels. Vikram Seth's The Golden Gate (1986), composed of 590 Onegin stanzas, is a more recent example of the verse novel.[17]

Experience of intimacy
Both in 12th-century Japan and 15th-century Europe, prose fiction created intimate reading situations. On the other hand, verse epics, including the Odyssey and Aeneid, had been recited to select audiences, though this was a more intimate experience than the performance of plays in theaters. A new world of individualistic fashion, personal views, intimate feelings, secret anxieties, "conduct", and "gallantry" spread with novels and the associated prose-romance.

Length
The novel is today the longest genre of narrative prose fiction, followed by the novella. However, in the 17th century, critics saw the romance as of epic length and the novel as its short rival. A precise definition of the differences in length between these types of fiction, is, however, not possible. The philosopher and literary critic György Lukács argued that the requirement of length is connected with the notion that a novel should encompass the totality of life.[18]